Lol I wanna test this feature of Github
I guess if I actually have to provide an actual change to propose, get rid of cohorts. You could still have a copy of the original file for cohorts if Covid gets bad again but make another copy for a file without cohorts to make the function have less parameters but idk that'll probably mean u will have to rewrite ur entire code which is annoying as hell. You do already have cohorts to be set to a default value of "None" which I'm assuming is null in python but you seem to use it to access a dictionary which could have undefined behavior but idk how ur code works, I don't really know much about the discord api or the json module u use to store all the data. Again, this is knowledge from someone who has barely used Github and the discord api/json modules/python in general. C++ be cool tho.